requirements.txt
requests>=2.25.0
pytest>=6.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0

# Optional: for SVG support (requires Cairo to be installed first)
# cairosvg>=2.5.0
